Refactor host store tests
