+++ src/storage/mod.rs
@@ -23,6 +23,13 @@ use crate::{
 #[derive(Debug, Clone, Copy)]
 pub struct Offset(u64);
 
+impl Offset {
+    /// Creates an offset with a given value
+    pub fn new(offset: u64) -> Offset {
+        Offset(offset)
+    }
+}
+
 /// An identifier representing a value stored somewhere else
 pub struct Ident<I, T> {
     id: I,
@@ -45,7 +52,7 @@ impl<I, T> Copy for Ident<I, T> where I: Copy {}
 
 impl<I, T> Ident<I, T> {
     /// Creates a typed identifier
-    pub(crate) fn new(id: I) -> Self {
+    pub fn new(id: I) -> Self {
         Ident {
             id,
             _marker: PhantomData,
